pyproject.toml
[project]
name = "clinic"
version = "0.1.0"
description = ""
authors = [
    {name = "vedantgadekar",email = "vedant.gadekar@mindstix.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "marshmallow (>=4.1.1,<5.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "flask-marshmallow (>=1.3.0,<2.0.0)",
    "bcrypt (>=4.0)"
] 


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-flask (>=1.3.0,<2.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
